move TS specific stuff to //cli/compiler.rs
