+++ cli/compiler.rs
@@ -1,24 +1,136 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
+use crate::deno_dir::get_cache_filename;
 use crate::deno_dir::DenoDir;
 use crate::deno_dir::SourceFile;
+use crate::deno_dir::SourceFileFetcher;
 use crate::diagnostics::Diagnostic;
+use crate::fs as deno_fs;
 use crate::msg;
 use crate::resources;
 use crate::startup_data;
 use crate::state::*;
+use crate::version;
 use crate::worker::Worker;
 use deno::Buf;
 use deno::ErrBox;
+use deno::ModuleSpecifier;
 use futures::future::Either;
 use futures::Future;
 use futures::Stream;
+use ring;
+use std::fmt::Write;
+use std::fs;
+use std::path::Path;
+use std::path::PathBuf;
 use std::str;
 use std::sync::atomic::Ordering;
+use url::Url;
 
 /// Optional tuple which represents the state of the compiler
 /// configuration where the first is canonical name for the configuration file
 /// and a vector of the bytes of the contents of the configuration file.
 type CompilerConfig = Option<(String, Vec<u8>)>;
+
+/// Information associated with compiled file in cache.
+/// Includes source code path and state hash.
+/// version_hash is used to validate versions of the file
+/// and could be used to remove stale file in cache.
+pub struct CompiledFileMetadata {
+  pub source_path: String,
+  pub version_hash: String,
+}
+
+static SOURCE_PATH: &'static str = "source_path";
+static VERSION_HASH: &'static str = "version_hash";
+
+impl CompiledFileMetadata {
+  // TODO: use serde for deserialization
+  /// Get compiled file metadata from meta_path.
+  /// If operation failed or metadata file is corrupted,
+  /// return None.
+  pub fn load<P: AsRef<Path>>(meta_path: P) -> Option<CompiledFileMetadata> {
+    let meta_path = meta_path.as_ref();
+    let maybe_metadata_string = fs::read_to_string(meta_path).ok();
+    maybe_metadata_string.as_ref()?;
+    let maybe_metadata_json: serde_json::Result<serde_json::Value> =
+      serde_json::from_str(&maybe_metadata_string.unwrap());
+    if let Ok(metadata_json) = maybe_metadata_json {
+      let source_path = metadata_json[SOURCE_PATH].as_str().map(String::from);
+      let version_hash = metadata_json[VERSION_HASH].as_str().map(String::from);
+      if source_path.is_none() || version_hash.is_none() {
+        return None;
+      }
+      return Some(CompiledFileMetadata {
+        source_path: source_path.unwrap(),
+        version_hash: version_hash.unwrap(),
+      });
+    } else {
+      return None;
+    }
+  }
+
+  /// Save compiled file metadata to meta_path.
+  pub fn save<P: AsRef<Path>>(self: &Self, meta_path: P) {
+    let meta_path = meta_path.as_ref();
+    // Remove possibly existing stale meta file.
+    // May not exist. DON'T unwrap.
+    let _ = std::fs::remove_file(&meta_path);
+    let mut value_map = serde_json::map::Map::new();
+
+    value_map.insert(SOURCE_PATH.to_owned(), json!(&self.source_path));
+    value_map.insert(VERSION_HASH.to_string(), json!(&self.version_hash));
+
+    let _ = serde_json::to_string(&value_map).map(|s| {
+      let _ = deno_fs::write_file(meta_path, s, 0o666);
+    });
+  }
+}
+
+/// Try loading the compiled code and map from the cache.
+/// Also validate the state hash in case source code has changed.
+/// If the version hash does not match,
+/// try delete all these files and return an error.
+fn load_cache(
+  js_filename: &PathBuf,
+  map_filename: &PathBuf,
+  meta_filename: &PathBuf,
+  version_hash_to_validate: String,
+) -> Result<(Vec<u8>, Vec<u8>), std::io::Error> {
+  debug!(
+    "load_cache code: {} map: {} meta: {}",
+    js_filename.display(),
+    map_filename.display(),
+    meta_filename.display()
+  );
+  let read_output_code = fs::read(&js_filename)?;
+  let read_source_map = fs::read(&map_filename)?;
+
+  // Validate metadata.
+  let mut should_remove_all = false;
+  if let Some(read_metadata) = CompiledFileMetadata::load(meta_filename) {
+    if read_metadata.version_hash != version_hash_to_validate {
+      debug!("load_cache metadata version hash does not match");
+      should_remove_all = true;
+    }
+  } else {
+    debug!("load_cache metadata load error");
+    should_remove_all = true;
+  }
+
+  // TODO: this step can be skipped now
+  // The old version is stale or some error occurred.
+  // Try removing all the cached files.
+  if should_remove_all {
+    let _ = fs::remove_file(&js_filename);
+    let _ = fs::remove_file(&map_filename);
+    let _ = fs::remove_file(&meta_filename);
+    // This will trigger a not-found error return
+    // and hint recompilation.
+    let _ = fs::read(&js_filename)?;
+  }
+
+  Ok((read_output_code, read_source_map))
+}
 
 /// Creates the JSON message send to compiler.ts's onmessage.
 fn req(
@@ -42,9 +154,38 @@ fn req(
   j.to_string().into_boxed_str().into_boxed_bytes()
 }
 
+fn gen_hash(v: Vec<&[u8]>) -> String {
+  let mut ctx = ring::digest::Context::new(&ring::digest::SHA1);
+  for src in v.iter() {
+    ctx.update(src);
+  }
+  let digest = ctx.finish();
+  let mut out = String::new();
+  // TODO There must be a better way to do this...
+  for byte in digest.as_ref() {
+    write!(&mut out, "{:02x}", byte).unwrap();
+  }
+  out
+}
+
+/// Emit a SHA1 hash based on source code, deno version and TS config.
+/// Used to check if a recompilation for source code is needed.
+pub fn source_code_version_hash(
+  source_code: &[u8],
+  version: &str,
+  config_hash: &str,
+) -> String {
+  gen_hash(vec![
+    source_code,
+    version.as_bytes(),
+    config_hash.as_bytes(),
+  ])
+}
+
 pub struct TsCompiler {
   pub deno_dir: DenoDir,
   pub config: CompilerConfig,
+  pub config_hash: String,
 }
 
 impl TsCompiler {
@@ -62,9 +203,15 @@ impl TsCompiler {
       _ => None,
     };
 
+    let config_hash = match &compiler_config {
+      Some((_, config)) => gen_hash(vec![&config.clone()[..]]),
+      None => gen_hash(vec![b""]),
+    };
+
     Self {
       deno_dir,
       config: compiler_config,
+      config_hash,
     }
   }
 
@@ -147,9 +294,7 @@ impl TsCompiler {
 
     if use_cache {
       // try to load cached version
-      if let Ok(compiled_module) =
-        self.deno_dir.get_compiled_source_file(&source_file)
-      {
+      if let Ok(compiled_module) = self.get_compiled_source_file(&source_file) {
         debug!(
           "found cached compiled module: {:?}",
           compiled_module.clone().filename
@@ -173,7 +318,7 @@ impl TsCompiler {
 
     let worker = TsCompiler::setup_worker(state.clone());
     let compiling_job = state.progress.add("Compile", &module_url.to_string());
-    let deno_dir_ = self.deno_dir.clone();
+    let state_ = state.clone();
 
     let resource = worker.state.resource.clone();
     let compiler_rid = resource.rid;
@@ -208,7 +353,7 @@ impl TsCompiler {
 
         Ok(())
       }).and_then(move |_| {
-        // TODO: must fetch compiled JS file from DENO_DIR, this can't file
+        // TODO: must fetch compiled JS file from DENO_DIR, this can't fail
         //  but if we get rid of gen/ and store JS files alongside TS files
         //  when we request local JS file it is fetched from original location on disk
         //  example:
@@ -242,7 +387,8 @@ impl TsCompiler {
         // We
         // still need to handle that JS and JSON files are compiled by TS compiler now (and they're
         // not used).
-        deno_dir_.get_compiled_source_file(&source_file_)
+        // TODO: this is bad
+        state_.ts_compiler.get_compiled_source_file(&source_file_)
           .map_err(|e| {
             // TODO(95th) Instead of panicking, We could translate this error to Diagnostic.
             panic!("{}", e)
@@ -260,6 +406,121 @@ impl TsCompiler {
       });
 
     Either::B(fut)
+  }
+
+  pub fn cache_paths(self: &Self, url: &Url) -> (PathBuf, PathBuf, PathBuf) {
+    let compiled_cache_filename = get_cache_filename(url);
+    (
+      compiled_cache_filename.with_extension("js"),
+      compiled_cache_filename.with_extension("js.map"),
+      compiled_cache_filename.with_extension("meta"),
+    )
+  }
+
+  pub fn get_compiled_source_file(
+    self: &Self,
+    source_file: &SourceFile,
+  ) -> Result<SourceFile, ErrBox> {
+    let compiled_cache_filename =
+      self.deno_dir.gen.join(get_cache_filename(&source_file.url));
+    let (
+      output_code_filename,
+      output_source_map_filename,
+      output_meta_filename,
+    ) = (
+      compiled_cache_filename.with_extension("js"),
+      compiled_cache_filename.with_extension("js.map"),
+      compiled_cache_filename.with_extension("meta"),
+    );
+
+    debug!(
+      "compiled filenames: {:?}, {:?}",
+      output_code_filename, output_source_map_filename
+    );
+
+    let version_hash_to_validate = source_code_version_hash(
+      &source_file.source_code,
+      version::DENO,
+      &self.config_hash,
+    );
+
+    // TODO: validation of hash should be done by compiler
+    let result = load_cache(
+      &output_code_filename,
+      &output_source_map_filename,
+      &output_meta_filename,
+      version_hash_to_validate,
+    );
+    debug!("load cache result: {:?}", result);
+    match result {
+      Err(err) => Err(err.into()),
+      Ok((output_code, source_map)) => {
+        let compiled_module = SourceFile {
+          url: source_file.url.clone(),
+          redirect_source_url: None,
+          filename: output_code_filename,
+          media_type: msg::MediaType::JavaScript,
+          source_code: output_code,
+          maybe_source_map: Some(source_map),
+          maybe_source_map_filename: Some(output_source_map_filename),
+        };
+
+        Ok(compiled_module)
+      }
+    }
+  }
+
+  pub fn cache_compiler_output(
+    self: &Self,
+    module_specifier: &ModuleSpecifier,
+    extension: &str,
+    contents: &str,
+  ) -> std::io::Result<()> {
+    let (js_cache_path, source_map_path, meta_data_path) =
+      self.cache_paths(module_specifier.as_url());
+    let (js_cache_path, source_map_path, meta_data_path) = (
+      self.deno_dir.gen.join(js_cache_path),
+      self.deno_dir.gen.join(source_map_path),
+      self.deno_dir.gen.join(meta_data_path),
+    );
+
+    // TODO: factor out to Enum
+    match extension {
+      ".map" => {
+        match source_map_path.parent() {
+          Some(ref parent) => fs::create_dir_all(parent),
+          None => unreachable!(),
+        }?;
+        fs::write(source_map_path, contents)?;
+      }
+      ".js" => {
+        let source_file = self
+          .deno_dir
+          .fetch_source_file(&module_specifier, true, true)
+          .expect("Source file not found");
+
+        match js_cache_path.parent() {
+          Some(ref parent) => fs::create_dir_all(parent),
+          None => unreachable!(),
+        }?;
+        fs::write(js_cache_path, contents)?;
+
+        // save .meta file
+        let version_hash = source_code_version_hash(
+          &source_file.source_code,
+          version::DENO,
+          &self.config_hash,
+        );
+        let compiled_file_metadata = CompiledFileMetadata {
+          source_path: source_file.filename.to_str().unwrap().to_string(),
+          version_hash,
+        };
+        compiled_file_metadata.save(&meta_data_path);
+      }
+      _ => unreachable!(),
+    }
+
+    Ok(())
   }
 }
